pyproject.toml
[project]
name = "forest-take-home"
version = "0.1.0"
description = "End-to-end data pipeline for London city data — Forest eBikes take-home"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.12.0",
    "responses>=0.25.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-requests>=2.31.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
